@@ -2,12 +2,15 @@
 plugins {
     id 'com.android.application'
     id 'org.jetbrains.kotlin.android'
-    id 'kotlin-android-extensions'
     id 'kotlin-kapt'
+    id 'com.google.dagger.hilt.android'
+
 }
 
 android {
-    compileSdk 32
+
+    namespace 'com.helpet'
+    compileSdk 33
 
     viewBinding {
         enabled true
@@ -19,12 +22,11 @@ android {
     defaultConfig {
         applicationId "com.helpet"
         minSdk 21
-        targetSdk 32
+        targetSdk 33
         versionCode 1
         versionName "1.0"
-
+        vectorDrawables.useSupportLibrary = true
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-
         multiDexEnabled true
     }
 
@@ -42,9 +44,6 @@ android {
         jvmTarget = '1.8'
     }
 
-    buildFeatures{
-        viewBinding true
-    }
 
 }
 
@@ -52,24 +51,25 @@ android {
 dependencies {
     //탭레이아웃
     //noinspection GradleCompatible
-    implementation 'com.android.support:design:28.0.0'
-    implementation 'androidx.core:core-ktx:1.7.0'
-    implementation 'androidx.appcompat:appcompat:1.5.1'
-    implementation 'com.google.android.material:material:1.7.0'
+    implementation 'androidx.core:core-ktx:1.10.1'
+    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'com.google.android.material:material:1.9.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
     implementation fileTree(dir: 'src\\main\\jniLibs', include: ['*.aar', '*.jar'], exclude: [])
     implementation 'com.google.android.gms:play-services-maps:18.1.0'
-    implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
-    implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.6.0'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.6.0'
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.4'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
 
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
-    implementation 'androidx.navigation:navigation-fragment:2.3.2'
-    implementation 'androidx.navigation:navigation-ui:2.3.2'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+    implementation 'androidx.navigation:navigation-fragment:2.6.0'
+    implementation 'androidx.navigation:navigation-ui:2.6.0'
     //tedpermission
     implementation 'io.github.ParkSangGwon:tedpermission:2.3.0'//권한 요청을 위한 팝업을 쉽게 구성하기 위해
-
+    // Coroutine Lifecycle Scopes
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.6.1"
+    implementation "androidx.lifecycle:lifecycle-runtime-ktx:2.6.1"
 
     // Retrofit
     implementation 'com.squareup.retrofit2:retrofit:2.9.0'
@@ -83,11 +83,30 @@ dependencies {
     implementation 'com.squareup.okhttp3:logging-interceptor:4.9.1'
 
     // Coroutine
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.1'
-    implementation 'androidx.core:core-splashscreen:1.0.0-rc01'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.3'
+//    implementation 'androidx.core:core-splashscreen:1.1.0-alpha01'
 
     //naver
     implementation 'com.naver.maps:map-sdk:3.14.0'
 
     implementation 'com.opencsv:opencsv:4.4'
+    implementation 'com.android.support:multidex:1.0.3'
+
+    //csv 공공데이터
+    implementation 'com.opencsv:opencsv:5.5.2'
+
+    implementation 'com.github.bumptech.glide:glide:4.12.0' // 최신 버전 사용 가능
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.12.0' // 최신 버전 사용 가능
+
+    //multimodule
+    implementation project(':domain')
+    implementation project(':data')
+
+    //hilt
+    implementation "com.google.dagger:hilt-android:2.44"
+    kapt "com.google.dagger:hilt-compiler:2.44"
+}
+
+kapt {
+    correctErrorTypes true
 }
