@@ -4,7 +4,7 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import com.helpet.R
-import kotlinx.android.synthetic.main.activity_choice_pet_species.*
+import com.helpet.databinding.ActivityChoicePetSpeciesBinding
 
 class ChoicePetSpecies : AppCompatActivity() {
 
@@ -13,24 +13,26 @@ class ChoicePetSpecies : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_choice_pet_species)
+
+        val binding = ActivityChoicePetSpeciesBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
 
-        back.setOnClickListener {
+        binding.back.setOnClickListener {
             val intent = Intent(this, VectorChoicePet::class.java)
             startActivity(intent)
         }
 
-        petSpeciesDog.setOnClickListener {
-            petSpeciesDog.setImageResource(R.drawable.choicedog)
+        binding.petSpeciesDog.setOnClickListener {
+            binding.petSpeciesDog.setImageResource(R.drawable.choicedog)
             petSpecies = "강아지"
             val intent = Intent(this, VectorCamera::class.java)
             intent.putExtra("petSpecies", petSpecies)
             startActivity(intent)
 
         }
-        petSpeciesCat.setOnClickListener {
-            petSpeciesDog.setImageResource(R.drawable.choicecat)
+        binding.petSpeciesCat.setOnClickListener {
+            binding.petSpeciesDog.setImageResource(R.drawable.choicecat)
             petSpecies = "고양이"
             val intent = Intent(this, VectorCamera::class.java)
             intent.putExtra("petSpecies", petSpecies)
