@@ -2,15 +2,15 @@ package com.helpet.Hospital
 
 import android.content.Intent
 import android.content.pm.PackageManager
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Base64
 import android.util.Log
+import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import com.google.android.material.tabs.TabLayout
 import com.helpet.R
+import com.helpet.databinding.ActivityHospitalBinding
 import com.helpet.vector.HomeActivity
-import kotlinx.android.synthetic.main.activity_hospital.*
 import java.security.MessageDigest
 import java.security.NoSuchAlgorithmException
 
@@ -20,7 +20,9 @@ class HospitalActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_hospital)
+
+        val binding = ActivityHospitalBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
 
         /* 키 해시 얻기*/try {
@@ -37,8 +39,8 @@ class HospitalActivity : AppCompatActivity() {
         }
 
 
-        back.setOnClickListener {
-            val intent= Intent(this, HomeActivity::class.java)
+        binding.back.setOnClickListener {
+            val intent = Intent(this, HomeActivity::class.java)
             startActivity(intent)
         }
 
@@ -47,22 +49,22 @@ class HospitalActivity : AppCompatActivity() {
 
         supportFragmentManager.beginTransaction().add(R.id.hospitalLayout, fragment0!!).commit()
 
-        tabs.addOnTabSelectedListener(object : TabLayout.OnTabSelectedListener {
+        binding.tabs.addOnTabSelectedListener(object : TabLayout.OnTabSelectedListener {
             override fun onTabSelected(tab: TabLayout.Tab) {
                 val position = tab.position
                 var selected: Fragment? = null
                 if (position == 0) {
                     selected = fragment0
-                }
-                else if (position == 1) {
+                } else if (position == 1) {
                     selected = fragment1
                 }
-                supportFragmentManager.beginTransaction().replace(R.id.hospitalLayout, selected!!).commit()
+                supportFragmentManager.beginTransaction().replace(R.id.hospitalLayout, selected!!)
+                    .commit()
             }
 
             override fun onTabUnselected(tab: TabLayout.Tab) {}
             override fun onTabReselected(tab: TabLayout.Tab) {}
         })
-    }
 
+    }
 }
