@@ -4,13 +4,12 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.Handler
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import com.helpet.vector.HomeActivity
 
 class TitleActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         setTheme(R.style.MyApplication)
-        installSplashScreen()
+//        installSplashScreen()
         super.onCreate(savedInstanceState)
 
         val intent= Intent(this, MainActivity::class.java)
