@@ -14,8 +14,10 @@ data class CalendarPlanResultDTO(
 data class CalendarPlanDTO(
     @SerializedName("cal_idx")
     val cal_idx: Int,
-    @SerializedName("date")
-    val date: String,
+    @SerializedName("startdate")
+    val startdate: String,
+    @SerializedName("enddate")
+    val enddate: String,
     @SerializedName("title")
     val title: String,
     @SerializedName("description")
@@ -32,6 +34,26 @@ data class MonthlyScheduleDTO(
     val userId : String,
     @SerializedName("result")
     val result: List<CalendarPlanDTO>
+)
+
+//월별 수정 DTO
+data class UpdateScheduleDTO(
+    @SerializedName("status")
+    val status: Boolean,
+//    @SerializedName("title")
+//    val title : String,
+//    @SerializedName("description")
+//    val description : String
+)
+
+//월별 삭제 DTO
+data class DeleteScheduleDTO(
+    @SerializedName("status")
+    val status: String,
+    @SerializedName("cal_idx")
+    val cal_idx : String,
+    @SerializedName("message")
+    val message : String
 )
 
 //data class MonthlyScheduleresponseDTO(
