@@ -1,9 +1,12 @@
 package com.helpet.calendar
 
 import retrofit2.Call
+import retrofit2.http.DELETE
 import retrofit2.http.Field
 import retrofit2.http.FormUrlEncoded
 import retrofit2.http.POST
+import retrofit2.http.PUT
+import retrofit2.http.Path
 
 //처음 접속 후 월별 일정 가져오기
 interface CalendarDateService {
@@ -30,11 +33,34 @@ interface CalendarService {
     @FormUrlEncoded
     @POST("calendar/add")
     fun CalendarResult(
-        @Field("date") date: String?,
+        @Field("startdate") startdate: String?,
+        @Field("enddate") enddate: String?,
         @Field("userId") userId: String,
         @Field("title") title: String,
         @Field("description") description: String
     ): Call<CalendarPlanResultDTO?>
 }
+
+
+//캘린더 수정
+interface CalendarEditService{
+    @FormUrlEncoded
+    @PUT("calendar/update")
+    fun CalendarUpdate(
+        @Field("cal_idx") cal_idx : Int?,
+        @Field("userId") userId: String,
+        @Field("startdate") startdate: String?,
+        @Field("enddate") enddate: String?,
+        @Field("title") title: String,
+        @Field("description") description: String
+    ) :Call<UpdateScheduleDTO>
+}
+//일정 삭제하기
+interface CalendarDeleteService{
+    @DELETE("calendar/delete/{cal_idx}")
+    fun calendarDelete(@Path("cal_idx") cal_idx: Int?): Call<DeleteScheduleDTO>
+}
+
+
 
 
