@@ -19,19 +19,18 @@ import androidx.annotation.LongDef
 import androidx.annotation.RequiresApi
 import androidx.core.view.isVisible
 import com.helpet.databinding.ActivityVectorCameraBinding
-import kotlinx.android.synthetic.main.activity_vector_camera.*
-import kotlinx.android.synthetic.main.activity_vector_result.*
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.MultipartBody
 import okhttp3.RequestBody
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
+import java.io.File
 import java.text.SimpleDateFormat
 import kotlin.concurrent.thread
 
 class VectorCamera : BaseActivity() {
-    var name = ""
+    var name :List<String> = listOf()
     var symptomProbability = 0.0
     var asymptomaticProbability = 0.0
     var vectContent = ""
@@ -41,16 +40,21 @@ class VectorCamera : BaseActivity() {
     val REQ_CAMERA=11
     val CROP_PICTURE = 2
 
+    private lateinit var binding : ActivityVectorCameraBinding
 
-    val binding by lazy { ActivityVectorCameraBinding.inflate(LayoutInflater.from(applicationContext)) }
+    private val diseaseList = mutableListOf<DiseaseName>()
+
+
     @RequiresApi(Build.VERSION_CODES.M)
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(binding.root)
-        buttonVector.isEnabled=false
-        buttonVector.backgroundTintList = ColorStateList.valueOf(Color.LTGRAY) // 회색으로 설정
 
-        camback.setOnClickListener {
+        binding = ActivityVectorCameraBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        binding.buttonVector.isEnabled=false
+        binding.buttonVector.backgroundTintList = ColorStateList.valueOf(Color.LTGRAY) // 회색으로 설정
+
+        binding.camback.setOnClickListener {
             val intent = Intent(this, VectorChoicePet::class.java)
             startActivity(intent)
             finish()
@@ -63,13 +67,6 @@ class VectorCamera : BaseActivity() {
     fun initViews(){
         binding.cameraBtn.setOnClickListener {
             requestPermissions(arrayOf(Manifest.permission.CAMERA),PERM_CAMERA)
-            camTitle.text="촬영 완료"
-            sub1.text="아래의 진단 시작 버튼을\n누르면 진단이 시작됩니다."
-            sub2.isVisible=false
-            sub3.isVisible=false
-            camSubTitle.isVisible=false
-            buttonVector.backgroundTintList = ColorStateList.valueOf(Color.parseColor("#FD9374")) // 오렌지색으로 설정
-
         }
     }
 
@@ -150,6 +147,7 @@ class VectorCamera : BaseActivity() {
         }
     }
 
+
     override fun permissionDenied(requestCode: Int) {
         when(requestCode){
             PERM_STORAGE->{
@@ -186,27 +184,46 @@ class VectorCamera : BaseActivity() {
                     startActivityForResult(intent, REQ_CAMERA);
                 }
                 REQ_CAMERA ->{
-                    buttonVector.isEnabled=true
-                    buttonVector.setBackgroundColor(Color.parseColor("#FD9374"))
+                    binding.buttonVector.isEnabled=true
+                    binding.buttonVector.setBackgroundColor(Color.parseColor("#FD9374"))
                     realUri?.let { uri ->
                         var bitmap: Bitmap? = null
                         //카메라에서 찍은 사진을 비트맵으로 변환
                         bitmap = MediaStore.Images.Media.getBitmap(contentResolver, realUri)
                         //이미지뷰에 이미지 로딩
                         binding.cameraBtn.setImageBitmap(bitmap)
+                        binding.camTitle.text="촬영 완료"
+                        binding.sub1.text="아래의 진단 시작 버튼을\n누르면 진단이 시작됩니다."
+                        binding.sub2.isVisible=false
+                        binding.sub3.isVisible=false
+                        binding.camSubTitle.isVisible=false
+                        binding.buttonVector.backgroundTintList = ColorStateList.valueOf(Color.parseColor("#FD9374")) // 오렌지색으로 설정
+                        binding.ReCam.isVisible = true
                     }
 
 
-                    buttonVector.setOnClickListener {
+                    binding.buttonVector.setOnClickListener {
                         var bitmap = MediaStore.Images.Media.getBitmap(contentResolver, realUri)
 
                         UpdatePhoto(SerialBitmap.translate(bitmap),this)
                         Log.d("사진" , SerialBitmap.translate(bitmap).toString())
-                        buttonVector.isVisible=false
-                        camSubLayout.isVisible=false
-                        loadingLayout.isVisible=true
-                        vectorProgress.isIndeterminate = true
-                        camTitle.text = "진단 중"
+                        binding.buttonVector.isVisible=false
+                        binding.camSubLayout.isVisible=false
+                        binding.loadingLayout.isVisible=true
+                        binding.vectorProgress.isIndeterminate = true
+                        binding.camTitle.text = "진단 중"
+                    }
+                    binding.ReCam.setOnClickListener {
+                        // 기존 이미지 파일 삭제
+                        if (realUri != null) {
+                            val file = File(getRealPathFromURI(realUri!!))
+                            if (file.exists()) {
+                                file.delete()
+                            }
+                        }
+
+                        // 다시 카메라 열기
+                        openCamera()
                     }
                 }
             }
@@ -227,6 +244,7 @@ private val server2 = RetrofitApi.retrofit.create(catVectorService::class.java)
         val sharedPreferences = getSharedPreferences("MyPrefs", Context.MODE_PRIVATE)
         // 유저아이디 데이터 읽기
         val value = sharedPreferences?.getString("userId", "null")
+        val namepet = intent.getStringExtra("namepet")
 
         val fileBody = RequestBody.create("image/*".toMediaTypeOrNull(), byteArray)
         val multipartBody: MultipartBody.Part? =
@@ -236,20 +254,25 @@ private val server2 = RetrofitApi.retrofit.create(catVectorService::class.java)
         //만약 강아지 눈진단이면 if문 활용
 
         if (petSpecies == "강아지"){
-            server.vectorResult(multipartBody!!).enqueue(object : Callback<ResponseDto?> {
-                override fun onResponse(call: Call<ResponseDto?>?, response: Response<ResponseDto?>) {
+            server.vectorResult(multipartBody!!).enqueue(object : Callback<ResultVectDTO?> {
+                override fun onResponse(call: Call<ResultVectDTO?>, response: Response<ResultVectDTO?>) {
 //                Toast.makeText(context, "File Uploaded Successfully...", Toast.LENGTH_LONG).show();
                     Log.d("레트로핏 결과2", "" + response.body().toString())
 
+                    for (diseaseName in response.body()?.diseaseNames!!) {
+                        diseaseList.add(DiseaseName(diseaseName))
+                    }
 
-                    val name= response.body()?.name!!
-                    val asymptomaticProbability= response.body()?.asymptomaticProbability!!
-                    val symptomProbability=response.body()?.symptomProbability!!
-                    val vectContent = response.body()?.vectContent!!
+
+                    name= response.body()?.diseaseNames!!
+                    asymptomaticProbability= response.body()?.asymptomaticProbability!!
+                    symptomProbability=response.body()?.symptomProbability!!
+//                    val vectContent = response.body()?.vectContent!!
                     // 다른 액티비티로 intent
                     val intent = Intent(context, VectorResult::class.java)
                     // 인텐트에 데이터 추가
-                    intent.putExtra("name", name)
+                    intent.putExtra("namepet", namepet )
+                    intent.putParcelableArrayListExtra("name" ,ArrayList(diseaseList))
                     intent.putExtra("symptomProbability",symptomProbability)
                     intent.putExtra("asymptomaticProbability",asymptomaticProbability )
                     intent.putExtra("vectImg",image)
@@ -257,9 +280,10 @@ private val server2 = RetrofitApi.retrofit.create(catVectorService::class.java)
                     intent.putExtra("value", value)
                     // 액티비티 시작
                     context.startActivity(intent)
+                    finish()
                 }
 
-                override fun onFailure(call: Call<ResponseDto?>?, t: Throwable) {
+                override fun onFailure(call: Call<ResultVectDTO?>, t: Throwable) {
 //                    Toast.makeText(context, "통신 실패", Toast.LENGTH_SHORT).show()
                     Log.d("에러", t.message!!)
                 }
@@ -270,31 +294,34 @@ private val server2 = RetrofitApi.retrofit.create(catVectorService::class.java)
 
         else if (petSpecies == "고양이"){
 
-            server2.catvectorResult(multipartBody!!).enqueue(object : Callback<ResponseDto?> {
-                override fun onResponse(call: Call<ResponseDto?>?, response: Response<ResponseDto?>) {
+            server2.catvectorResult(multipartBody!!).enqueue(object : Callback<ResultVectDTO?> {
+                override fun onResponse(call: Call<ResultVectDTO?>, response: Response<ResultVectDTO?>) {
 //                Toast.makeText(context, "File Uploaded Successfully...", Toast.LENGTH_LONG).show();
-                    Log.d("레트로핏 결과2", "" + response.body().toString())
 
+                    for (diseaseName in response.body()?.diseaseNames!!) {
+                        diseaseList.add(DiseaseName(diseaseName))
+                    }
 
-                    name= response.body()?.name!!
+                    name= response.body()?.diseaseNames!!
                     asymptomaticProbability= response.body()?.asymptomaticProbability!!
                     symptomProbability=response.body()?.symptomProbability!!
+//                    val vectContent = response.body()?.vectContent!!
                     // 다른 액티비티로 intent
                     val intent = Intent(context, VectorResult::class.java)
                     // 인텐트에 데이터 추가
-                    intent.putExtra("name", name)
+                    intent.putExtra("namepet", namepet )
+                    intent.putParcelableArrayListExtra("name" ,ArrayList(diseaseList))
                     intent.putExtra("symptomProbability",symptomProbability)
                     intent.putExtra("asymptomaticProbability",asymptomaticProbability )
-                    intent.putExtra("vectImg",image )
+                    intent.putExtra("vectImg",image)
                     intent.putExtra("vectContent", vectContent)
                     intent.putExtra("value", value)
-
-
                     // 액티비티 시작
                     context.startActivity(intent)
+                    finish()
                 }
 
-                override fun onFailure(call: Call<ResponseDto?>?, t: Throwable) {
+                override fun onFailure(call: Call<ResultVectDTO?>, t: Throwable) {
                     Toast.makeText(context, "통신 실패", Toast.LENGTH_SHORT).show()
                     Log.d("에러", t.message!!)
                 }
