@@ -64,22 +64,40 @@ exports.petDelete = (req, res) => {
     var userId = req.body.userId;
     var petName = req.body.petName;
 
-    db.query('DELETE FROM pet WHERE petName = ? and userId = ?', [petName, userId], function (error, data) {
+    db.query('SELECT petIdx FROM pet WHERE petName = ? and userId = ?', [petName, userId], function (error, petData) {
         if (error) {
             return res.status(500).json({ error: error.message });
         }
-        res.json({
-            status: "success",
-            userId: userId,
-            petName: petName,
-        })
-        // if (db.state === 'connected') {
-        //     // 연결 종료
-        //     // db.end();
-        // }
-    })
-}
 
+        if (petData.length > 0) {
+            const petIdx = petData[0].petIdx;
+            db.query('DELETE FROM diag_pet WHERE petIdx = ?', [petIdx], function (error, data) {
+                if (error) {
+                    return res.status(500).json({ error: error.message });
+                }
+
+                db.query('DELETE FROM pet WHERE petName = ? and userId = ?', [petName, userId], function (error, data) {
+                    if (error) {
+                        return res.status(500).json({ error: error.message });
+                    }
+
+                    res.json({
+                        status: "success",
+                    });
+                });
+            });
+        } else {
+            db.query('DELETE FROM pet WHERE petName = ? and userId = ?', [petName, userId], function (error, data) {
+                if (error) {
+                    return res.status(500).json({ error: error.message });
+                }
+                res.json({
+                    status: "success",
+                });
+            });
+        }
+    });
+}
 //진단결과 리스트 조회 (진단리스트 홈화면)
 exports.petList = (req, res) => {
     var userId = req.body.userId;
@@ -93,36 +111,41 @@ exports.petList = (req, res) => {
             return res.status(404).json({ message: "No results found." });
         }
         res.json({
-            status: "success",
-            petName: petName,
             petAge: result[0].petAge,
             petBirth: result[0].petBirth,
+            petName: result[0].petName,
             result: result.map(item => ({
                 vectImg: item.vectImg,
                 vectDate: item.vectDate,
-                vectName: item.vectName,
+                vectName: item.vectName.split(","), 
                 vectProb: item.vectProb
-            }))
-        })
+            })),
+            status: "success"
+        });
         // if (db.state === 'connected') {
         //     // 연결 종료
         //     // db.end();
         // }
     })
-}
-
-//진단결과 저장
+}//진단결과 저장
 exports.petListSave = (req, res) => {
     const file = req.file;
 
     var userId = req.body.userId;
+    console.log("userId", userId)
     var petName = req.body.petName;
+    console.log("petName", petName)
     var vectDate = req.body.vectDate;
+    console.log("vectDate", vectDate)
     var vectName = req.body.vectName;
+    console.log("vectName", vectName)
     var vectProb = req.body.vectProb;
+    console.log("vectProb", vectProb)
     var vectContent = req.body.vectContent;
+    console.log("vectContent", vectContent)
 
     var imgName = file.filename;
+    console.log(imgName)
     const filePath = path.join(__dirname, '..', 'main', 'uploads', imgName);
     const imageBuffer = fs.readFileSync(filePath);
     let vectImg = Buffer.from(imageBuffer).toString('base64');
@@ -136,8 +159,11 @@ exports.petListSave = (req, res) => {
         }
         const petIdx = result[0].petIdx;
 
-        db.query('INSERT INTO diag_pet(diagIdx, petIdx, vectImg, vectDate, vectName, vectProb, vectContent) VALUES (?,?,?,?,?,?,?)', [null, petIdx, vectImg, vectDate, vectName, vectProb, vectContent], function (error, result) {
+        console.log("petIdx", petIdx)
+
+        db.query('INSERT INTO diag_pet(diagIdx, petIdx, vectImg, vectDate, vectName, vectProb, vectContent) VALUES (?,?,?,?,?,?,?)', [null, petIdx, vectImg, vectDate, JSON.stringify(vectName), vectProb, vectContent], function (error, result) {
             if (error) {
+                console.log("error", error.message)
                 return res.status(500).json({ error: error.message });
             }
             res.json({
